requirements.txt
streamlit>=1.28.0
exifread>=3.0.0
opencv-python>=4.8.0
pillow>=10.0.0
reportlab>=4.0.0
plotly>=5.17.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.28.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-image>=0.21.0
scikit-learn>=1.3.0 